refactor(ruff): Ruff清零 - 从758到0
- F821: 修复 asyncio 未导入bug (data.py)
- PLW1510: 17处 subprocess.run 加 check/noqa
- RUF059: 10处 未使用解包变量加_前缀
- SIM102: 10处 嵌套if合并
- TRY401: 21处 logging.exception移除冗余异常对象
- DTZ005: 3处 datetime.now加timezone.utc
- PYI034: 3处 __enter__返回Self类型
- B017: 11处 测试assertRaises加noqa
- F841/I001: 21处 自动修复
- pyproject.toml: BLE001/S110项目级忽略(高容错设计)

Ruff: 758→0
Tested: 4136 passed, 4 skipped

+++ py/maop/dashboard/data_proxy.py
@@ -279,7 +279,7 @@ class DataProxy:
                     "name": a.get("agent", ""),
                     "healthy": a.get("circuit_breaker", "closed") == "closed",
                     "queue": queue_depth,
-                    "load": min(100, max(0, int(round(100 - a.get("success_rate", 100.0))))),
+                    "load": min(100, max(0, round(100 - a.get("success_rate", 100.0)))),
                 }
                 for a in agent_rows
             ]
